VoidHaulerRework
Error with the counting is due to our bigrams not having the obviously correct amount of unigrams to compare against. With a bigram structured as (A,B) we must count all the occurences of A in the cleansed Bigram distribution to have the correct number to compare against. The probabilities now get correctly calculated (as well as it can before we can correct the counts.